+++ Content/Buffs/DizzyDebuff.cs
@@ -1,0 +1,25 @@
+﻿using PaperMarioItems.Common.NPCs;
+using PaperMarioItems.Common.Players;
+using Terraria;
+using Terraria.Localization;
+using Terraria.ModLoader;
+
+namespace PaperMarioItems.Content.Buffs
+{
+	public class DizzyDebuff : ModBuff
+	{
+        public override LocalizedText Description => base.Description;
+        public override void SetStaticDefaults()
+        {
+            Main.pvpBuff[Type] = true;
+        }
+        public override void Update(NPC npc, ref int buffIndex)
+        {
+            npc.GetGlobalNPC<PaperNPC>().dizzyDebuff = true;
+        }
+        public override void Update(Player player, ref int buffIndex)
+        {
+            player.GetModPlayer<PaperPlayer>().dizzyDebuff = true;
+        }
+    }
+}